Update ImageData and ImageBitmap sections
Move the ImageDataColorSettings definition to the ImageData section of the specification.

Update ImageDataSettings to use the existing ImageDataStorageType type, and rename the ImageDataStorageTypes of uint8 and uint16 to unorm8 and unorm16.

Change the default ImageDataSettings for createImageData and getImageData to be their current values of uint8 and srgb. This is preferable because it is more well-defined (there is no implicit selection of the "best matching" ImageDataStorageType) and because it preserves existing behavior.

Add text indicating that ImageBitmap will make its best effort to preserve the fidelity of what it is representing.