Indicate implementations should not limit exposed color spaces
Remove "Implementations should not expose color spaces that are unreasonble for the display." Add "Implementations should not limit the set of exposed color spaces based on the capabilities of the display."

The motivation for this is that applications should not be limited by the hardware they happen to be running on. A good example is an application to do photo cropping and zooming using a canvas or WebGL. Such an application will want to match the input image color space, and so it will need to use that (or a larger) color space as its working space.