Change colorEncoding to storageFormat.
Rename colorEncoding to storageFormat. This mirrors the names used in ImageData, and clarifies that this specifies the format primarily and the color encoding function only for some formats.

Update the sRGB colorEncoding format to match the function specified in section 4.1.8 of the OpenGL ES 3.0.3 specification (removing the explicit 8-bit references).